[Interfaces] (Changes after comments)
- Moved service interfaces closer to clients into ./pkg/interface/ folder.
- Deleted redundant controller interfaces.
