+++ internal/job/send-emails-job.go
@@ -1,9 +1,9 @@
 package job
 
 import (
+	"genesis-currency-api/pkg/interface/service"
 	"log"
 
-	"genesis-currency-api/internal/service"
 	"github.com/robfig/cron/v3"
 )
