+++ pkg/interface/service/currency-service.go
@@ -1,0 +1,9 @@
+package service
+
+import "genesis-currency-api/pkg/dto"
+
+type CurrencyService interface {
+	GetCurrencyInfo() dto.CurrencyInfoDTO
+	GetCurrencyRate() dto.CurrencyResponseDto
+	UpdateCurrencyRates() error
+}
