+++ pkg/interface/service/email-service.go
@@ -1,0 +1,5 @@
+package service
+
+type EmailService interface {
+	SendEmails() error
+}
